cardapp application
This Flutter project is a basic profile card application that demonstrates the use of widgets for UI design. It displays a personal profile containing a name, role, contact details, and styling with icons and layouts. This project is ideal for beginners learning Flutter’s Scaffold, Card, Column, Row, and Icon widgets.